Changes made while testing out OData. It is not working well with Swagger, so I abandoned it in favor simple REST. Fixes #130
