+++ src/Devantler.DataProduct.Generator/IncrementalGenerators/RestControllerGenerator.cs
@@ -6,7 +6,6 @@ using Devantler.Commons.CodeGen.CSharp.Model;
 using Devantler.Commons.CodeGen.Mapping.Avro;
 using Devantler.Commons.StringHelpers.Extensions;
 using Devantler.DataProduct.Configuration.Options;
-using Devantler.DataProduct.Configuration.Options.FeatureFlags;
 using Devantler.DataProduct.Generator.Extensions;
 using Devantler.DataProduct.Generator.Models;
 using Microsoft.CodeAnalysis;
@@ -17,7 +16,7 @@ namespace Devantler.DataProduct.Generator.IncrementalGenerators;
 /// A generator that generates REST API controllers in the data product.
 /// </summary>
 [Generator]
-public class CRUDBulkControllerGenerator : GeneratorBase
+public class CRUDControllerGenerator : GeneratorBase
 {
     /// <summary>
     /// Generates REST API controllers in the data product.
@@ -30,9 +29,6 @@ public class CRUDBulkControllerGenerator : GeneratorBase
         ImmutableArray<AdditionalFile> additionalFiles,
         DataProductOptions options)
     {
-        if (!options.FeatureFlags.EnableApis.Contains(ApiFeatureFlagValues.Rest) || !options.Apis.Rest.EnableBulkControllers)
-            return new Dictionary<string, string>();
-
         var schemaRegistryClient = options.SchemaRegistry.CreateSchemaRegistryClient();
         var rootSchema = schemaRegistryClient.GetSchema(options.SchemaRegistry.Schema.Subject, options.SchemaRegistry.Schema.Version);
 
@@ -45,15 +41,15 @@ public class CRUDBulkControllerGenerator : GeneratorBase
             string schemaIdType = schemaType is not null
                 ? avroSchemaParser.Parse(schemaType, Language.CSharp)
                 : "Guid";
-            var @class = new CSharpClass($"{schemaName.ToPlural()}Controller")
+            var @class = new CSharpClass($"{schemaName}Controller")
                 .AddImport(new CSharpUsing("AutoMapper"))
                 .AddImport(new CSharpUsing(NamespaceResolver.ResolveForType(compilation.GlobalNamespace, "ISchema")))
                 .AddImport(new CSharpUsing(NamespaceResolver.ResolveForType(compilation.GlobalNamespace, "IEntity")))
                 .AddImport(new CSharpUsing(NamespaceResolver.ResolveForType(compilation.GlobalNamespace, "DataStoreService")))
-                .SetNamespace(NamespaceResolver.ResolveForType(compilation.GlobalNamespace, "CRUDBulkController"))
+                .SetNamespace(NamespaceResolver.ResolveForType(compilation.GlobalNamespace, "RestController"))
                 .SetDocBlock(new CSharpDocBlock(
-                    $$"""A controller to handle REST API requests for a the <see cref="{{schemaName}}" /> schema."""))
-                .SetBaseClass(new CSharpClass($"CRUDBulkController<{schemaIdType}, {schemaName}>"));
+                    $$"""A controller to handle CRUD REST API requests for a the <see cref="{{schemaName}}" /> schema."""))
+                .SetBaseClass(new CSharpClass($"RestController<{schemaIdType}, {schemaName}>"));
 
             var constructor = new CSharpConstructor(@class.Name)
                 .SetDocBlock(new CSharpDocBlock($$"""Creates a new instance of <see cref="{{@class.Name}}" />"""));
