+++ src/Devantler.DataProduct/config.yaml
@@ -34,6 +34,10 @@ DataStore:
   Provider: Sqlite
   ConnectionString: Data Source=sqlite.db
 
+Apis:
+  Rest:
+    EnableBulkController: true
+
 CacheStore:
   Type: InMemory
