+++ tests/Devantler.DataProduct.Generator.Tests.Unit/IncrementalGenerators/RestBulkControllerGeneratorTests/GenerateTests.record-schema-primitive-types#RecordSchemaPrimitiveTypesController.g.verified.cs
