+++ tests/Devantler.DataProduct.Generator.Tests.Unit/IncrementalGenerators/RestBulkControllerGeneratorTests/TestCases.cs
@@ -1,4 +1,4 @@
-namespace Devantler.DataProduct.Generator.Tests.Unit.IncrementalGenerators.CRUDControllerGeneratorTests;
+namespace Devantler.DataProduct.Generator.Tests.Unit.IncrementalGenerators.BulkControllerGeneratorTests;
 
 public static class TestCases
 {
